refactor: reorganize notes folder structure
- Merge decrypt-commitment.ts and commitment.ts into single commitment.ts
- Dissolve note-utils.ts: assertValidRandom and getHash become Note static
  methods, legacy helpers become TransactNote static methods
- token-utils.ts uses shared formatToByteLength/bigintToHex from encoding.ts
- Remove unnecessary Uint8Array wrappers in address-utils.ts
- Update barrel exports and test imports
