+++ src/notes/index.ts
@@ -1,12 +1,10 @@
 export * from './address-utils'
 export * from './commitment'
-export * from './decrypt-commitment'
 export * from './definitions'
 export * from './token-utils'
-export * from './note-utils'
-export * from './wallet-info'
-export * from './memo'
 
+export { WalletInfo } from './wallet-info'
+export { Memo } from './memo'
 export { Note } from './note'
 export { ShieldNote } from './shield-note'
 export { UnshieldNote } from './unshield-note'
